requirements.txt
streamlit

numpy
pandas
matplotlib.pyplot